Refactor dials.reindex
Move useful functions to dials.util.reindex for reuse
