+++ src/dials/command_line/reindex.py
@@ -3,7 +3,6 @@
 
 from __future__ import annotations
 
-import copy
 import logging
 import os
 import sys
@@ -13,19 +12,21 @@ import numpy as np
 import iotbx.phil
 from cctbx import sgtbx
 from libtbx import Auto
-from rstbx.symmetry.constraints import parameter_reduction
 
 import dials.util
 from dials.algorithms.indexing.assign_indices import AssignIndicesGlobal
 from dials.algorithms.scaling.scaling_library import determine_best_unit_cell
-from dials.algorithms.symmetry.reindex_to_reference import (
-    determine_reindex_operator_against_reference,
-)
 from dials.array_family import flex
 from dials.util import Sorry, log
 from dials.util.filter_reflections import filtered_arrays_from_experiments_reflections
 from dials.util.options import ArgumentParser, reflections_and_experiments_from_files
 from dials.util.reference import intensities_from_reference_file
+from dials.util.reindex import (
+    change_of_basis_op_against_reference,
+    derive_change_of_basis_op,
+    reindex_experiments,
+    reindex_reflections,
+)
 from dials.util.version import dials_version
 
 logger = logging.getLogger("dials.command_line.reindex")
@@ -98,115 +99,6 @@ output {
 """,
     process_includes=True,
 )
-
-
-def derive_change_of_basis_op(from_hkl, to_hkl):
-
-    # exclude those reflections that we couldn't index
-    sel = (to_hkl != (0, 0, 0)) & (from_hkl != (0, 0, 0))
-    assert sel.count(True) >= 3  # need minimum of 3 equations ?
-    to_hkl = to_hkl.select(sel)
-    from_hkl = from_hkl.select(sel)
-
-    # for each miller index, solve a system of linear equations to find the
-    # change of basis operator
-    h, k, l = to_hkl.as_vec3_double().parts()
-
-    r = []
-    from scitbx.lstbx import normal_eqns
-
-    for i in range(3):
-        eqns = normal_eqns.linear_ls(3)
-        for index, hkl in zip((h, k, l)[i], from_hkl):
-            eqns.add_equation(
-                right_hand_side=index, design_matrix_row=flex.double(hkl), weight=1
-            )
-        eqns.solve()
-        r.extend(eqns.solution())
-
-    from scitbx import matrix
-    from scitbx.math import continued_fraction
-
-    denom = 12
-    r = [
-        int(denom * continued_fraction.from_real(r_, eps=1e-2).as_rational())
-        for r_ in r
-    ]
-    r = matrix.sqr(r).transpose()
-    # print (1/denom)*r
-
-    # now convert into a cctbx change_of_basis_op object
-    change_of_basis_op = sgtbx.change_of_basis_op(
-        sgtbx.rt_mx(sgtbx.rot_mx(r, denominator=denom))
-    ).inverse()
-    logger.info(f"discovered change_of_basis_op={change_of_basis_op}")
-
-    # sanity check that this is the right cb_op
-    assert (change_of_basis_op.apply(from_hkl) == to_hkl).count(False) == 0
-
-    return change_of_basis_op
-
-
-def reindex_experiments(experiments, cb_op, space_group=None):
-    reindexed_experiments = copy.deepcopy(experiments)
-
-    for crystal in reindexed_experiments.crystals():
-        cryst_reindexed = copy.deepcopy(crystal)
-        if space_group is not None:
-            # See also https://github.com/cctbx/cctbx_project/issues/424
-            cryst_reindexed.set_space_group(sgtbx.space_group("P 1"))
-            cryst_reindexed = cryst_reindexed.change_basis(cb_op)
-            cryst_reindexed.set_space_group(space_group)
-            S = parameter_reduction.symmetrize_reduce_enlarge(
-                cryst_reindexed.get_space_group()
-            )
-            S.set_orientation(cryst_reindexed.get_B())
-            S.symmetrize()
-            # Cache the scan-varying A matrices if applicable as these get lost
-            # when we call crystal.set_B()
-            A_varying = [
-                cryst_reindexed.get_A_at_scan_point(i)
-                for i in range(cryst_reindexed.num_scan_points)
-            ]
-            # Update the symmetrized B matrix
-            cryst_reindexed.set_B(S.orientation.reciprocal_matrix())
-            # Reapply the scan-varying A matrices
-            cryst_reindexed.set_A_at_scan_points(A_varying)
-        else:
-            cryst_reindexed = cryst_reindexed.change_basis(cb_op)
-        crystal.update(cryst_reindexed)
-
-    return reindexed_experiments
-
-
-def change_of_basis_op_against_reference(
-    experiments, reflections, reference_miller_set
-):
-    # Set some flags to allow filtering, if wanting to reindex against
-    # reference with data that has not yet been through integration
-    for table in reflections:
-        if table.get_flags(table.flags.integrated_sum).count(True) == 0:
-            assert (
-                "intensity.sum.value" in table
-            ), "No 'intensity.sum.value' in reflections"
-            table.set_flags(flex.bool(table.size(), True), table.flags.integrated_sum)
-    # Make miller array of the datasets
-    filter_logger = logging.getLogger("dials.util.filter_reflections")
-    filter_logger.disabled = True
-    try:
-        test_miller_sets = filtered_arrays_from_experiments_reflections(
-            experiments, reflections, partiality_threshold=0.4
-        )
-    except ValueError:
-        raise Sorry("No reflections remain after filtering the test dataset")
-    test_miller_set = test_miller_sets[0]
-    for d in test_miller_sets[1:]:
-        test_miller_set = test_miller_set.concatenate(d)
-
-    change_of_basis_op = determine_reindex_operator_against_reference(
-        test_miller_set, reference_miller_set
-    )
-    return change_of_basis_op
 
 
 @dials.util.show_mail_handle_errors()
@@ -309,9 +201,12 @@ experiments file must also be specified with the option: reference.experiments= 
         except ValueError:
             raise Sorry("No reflections remain after filtering the reference dataset")
 
-        change_of_basis_op = change_of_basis_op_against_reference(
-            experiments, reflections, reference_miller_set
-        )
+        try:
+            change_of_basis_op = change_of_basis_op_against_reference(
+                experiments, reflections, reference_miller_set
+            )
+        except ValueError:
+            raise Sorry("No reflections remain after filtering the test dataset")
 
     elif params.reference.file:
 
@@ -397,28 +292,9 @@ experiments file must also be specified with the option: reference.experiments= 
         experiments.as_file(params.output.experiments)
 
     if len(reflections):
-        reflections = flex.reflection_table.concat(reflections)
-
-        miller_indices = reflections["miller_index"]
-
-        if params.hkl_offset is not None:
-            h, k, l = miller_indices.as_vec3_double().parts()
-            h += params.hkl_offset[0]
-            k += params.hkl_offset[1]
-            l += params.hkl_offset[2]
-            miller_indices = flex.miller_index(h.iround(), k.iround(), l.iround())
-        non_integral_indices = change_of_basis_op.apply_results_in_non_integral_indices(
-            miller_indices
+        reflections = reindex_reflections(
+            reflections, change_of_basis_op, params.hkl_offset
         )
-        if non_integral_indices.size() > 0:
-            logger.info(
-                f"Removing {non_integral_indices.size()}/{miller_indices.size()} reflections (change of basis results in non-integral indices)"
-            )
-        sel = flex.bool(miller_indices.size(), True)
-        sel.set_selected(non_integral_indices, False)
-        miller_indices_reindexed = change_of_basis_op.apply(miller_indices.select(sel))
-        reflections["miller_index"].set_selected(sel, miller_indices_reindexed)
-        reflections["miller_index"].set_selected(~sel, (0, 0, 0))
 
         logger.info(f"Saving reindexed reflections to {params.output.reflections}")
         reflections.as_file(params.output.reflections)
